Fix uploading tempfiles to S3 using aws-sdk
For some reason aws-sdk is throwing timeout errors, and this was a known
problem (https://github.com/aws/aws-sdk-ruby/issues/241), which is
currently supposed to be fixed. Anyway, for some reason this happens
only with Tempfiles that were changed by ImageMagick, I tried calling
\#fsync but it didn't work. So, to avoid this bug, we simply reopen the
tempfile to get a fresh file descriptor again.