Use Vips::Image#thumbnail_image
Vips::Image#thumbnail_image uses `vips_thumbnail()`, which has common
features for generating thumbnails, and according to the VIPS author
it has better quality and is also faster than the regular
`vips_resize()`.

This allows us to remove all the logic for calculating dimensions,
because `vips_thumbnail()` does the right thing.

This also autorotates the images on processing, which is what we'll
apply to MiniMagick as well.