[minimagick] Suggest passing transparent color as a Symbol
To make a distinction that "transparent" is not a valid ImageMagick
color, we suggest using a Symbol, and reserve strings only for valid
ImageMagick color.

We also fix a typo where we used :color option instead of :background.

We also remove the "TRANSPARENT" constant, as any RGB values with the
alpha of 0 will be transparent, so this is not one and only transparent
color, it's just transparent white.