Make the pipeline object accessible to processors
In certain cases processors need to know the source/destination format,
e.g. when they need to perform different operations for different
formats. So we refactor the pipeline class and initialize processors
with it.

With this we also create a separate Builder class, whose sole purpose is
to build options and call the pipeline with those options. This is an
improvement, because previously the Pipeline class was also acting like
a builder by including Chainable. This was making the Pipeline class
difficult to change, as Chainable#method_missing was swallowing any
typos. It was also making it impossible to refactor, because we would
risk the new methods clashing with VIPS/ImageMagick operation names.