Revert "Add back caching gems on Travis CI"
This apparently didn't force Travis CI to rebuild the Bundler cache, it
just reused the old one it had saved, so tests for Ruby 2.5 are failing
again due to the RubyGems/Bundler/Rake bug.

Since the test run time is dominated by installing libvips and
ImageMagick from source (which we have to if we want versions that are
new enough), I don't mind the `bundle install` adding extra 5% to the
build time.