[minimagick] Allow passing nil to add ImageMagick options that don't have a value
In addition to `strip: true`, the user might also find it logical to
pass `strip: nil` in the loader/saver options to do the same thing. For
example, CarrierWave::MiniMagick takes that approach.

In order to make the API more redundant and increase compatibility with
CarrierWave::MiniMagick interface, we allow passing `strip: nil`.