[core] Add #apply for applying a list of operations
Both CarrierWave and ActiveStorage give the ability for the user to
provide a hash of operations that should be applied. Since the plan is
to make CarrierWave and ActiveStorage depend on ImageProcessing, I think
it's nice for ImageProcessing to provide this functionality out of the
box.

In addition to the hash of operations, we make sure we accept an array
of operations as well, so that users can repeat the same operation
multiple times.

This removes the ability for ImageProcessing::MiniMagick to accept
"false" for "+" operators in loader/saver options, but this wasn't
documented anyway. If this feature is requested, we'll find another way.