Add back official GraphicsMagick support
In 4437199987dd65b8c139e2ad48ddde7d821d075a I decided to drop
GraphicsMagick support, because it made the MiniMagick test suite
convoluted and slow, as it had to run once for ImageMagick and once for
GraphicsMagick.

However, we can solve this issue by running GraphicsMagick tests just on
Travis CI. And it's really easy to skip assertions that are only
valid on ImageMagick.

We also add documentation on how to enable GraphicsMagick support.

Closes #36