[vips] Fix auto-rotation bug on libvips 8.9.0
When performing resizing with loader options, the following gets called
internally:

  Vips::Image.new_from_file("/path/to/source.jpg")
    .autorot
    .thumbnail_image(400, 400)
    .write_to_file("/path/to/destination.jpg")

Prior to libvips 8.9.0, this code would produce a correctly auto-rotated
image. However, in libvips 8.9.0, the auto-rotation from `autorot` and
the default auto-rotation performed by `thumbnail_image` cancel each
other out, leaving the image rotated.

We fix this by disabling auto-rotation for the `thumbnail` operation, so
that the only auto-rotation done is by `autorot`. In the resize-on-load
case we leave `thumbnail` to auto-rotate, as in that case `autorot` is
not called beforehand.

Closes #64